requirements.txt
Django==3.1.2
django-environ==0.4.5
djangorestframework==3.12.1appdirs==1.4.3